+++ src/anime_bridge/__init__.py
@@ -1,3 +1,3 @@
 """Anime Bridge public package metadata."""
 
-__version__ = "0.7.0.dev0"
+__version__ = "0.8.0.dev0"